rpm: improve packaging

+++ pkg-scripts/preinstall.sh
@@ -11,4 +11,4 @@ if [ -z `getent passwd frontman` ]; then
 fi
 
 # remove deprecated sysctl setting
-rm -f /etc/sysctl.d/50-ping_group_range.conf
+test -e /etc/sysctl.d/50-ping_group_range.conf && rm -f /etc/sysctl.d/50-ping_group_range.conf
